+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to BSDBuild will be documented in this file. This project ad
 
 ### Added
 
+- [**build.prog.mk**](https://bsdbuild.hypertriton.com/man5/build.prog.mk) & [**build.lib.mk**](https://bsdbuild.hypertriton.com/man5/build.lib.mk): Add support for compiling to wasm (WebAssembly) with Emscripten (https://emscripten.org).
 - [**build.prog.mk**](https://bsdbuild.hypertriton.com/man5/build.prog.mk) & [**build.lib.mk**](https://bsdbuild.hypertriton.com/man5/build.lib.mk): Add support for the Ada language. Introduce `$ADA`, `$ADABIND`, `$ADALINK` and `$LINKER_TYPE`. Introduce `${PROG_BUNDLE}`, `${LIB_BUNDLE}` and `gen-bundle.pl` for generating platform-specific application bundles (currently "OSX" or "iOS").
 - [**build.www.mk**](https://bsdbuild.hypertriton.com/man5/build.www.mk): Introduce the mlproc(1) multilanguage preprocessor. Define `${MLPROC}`, `${MLPROCFLAGS}`, `${MINIFIER}`, `${MINIFIERFLAGS}` and `${MINIFIERFLAGSCSS}`.
 - [**mkconfigure**](https://bsdbuild.hypertriton.com/man1/mkconfigure): New directives: `Ada_option()`, `Ada_bflag()`, `Hdefine_if()`. Export LDFLAGS to Makefile.config. Fix `Ld_Option()`. New directive `Pkgconfig_Module()` for integrated pkg-config module generation. Add `$PKGCONFIG_LIBDIR` setting to simplify the installation of .pc modules. Introduce `--keep-conftest` configure option (preserve output test files).